requirements.txt
numpy
scipy
networkx
matplotlib
pillow
torch
pandas
pyreadstat
jupyter
wandb
six
torch-geometric